refactor: rename apis
